% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_wrap.R
\name{addjust_beta}
\alias{addjust_beta}
\title{Adjust Beta Values to Lie Between 0 and 1}
\usage{
addjust_beta(number)
}
\arguments{
\item{number}{A numeric value to adjust between 0 and 1.}
}
\value{
Returns the adjusted number constrained between 0 and 1.
}
\description{
This function adjusts the input number to ensure it falls between 0 and 1. If the number 
is less than 0, it returns 0, and if greater than 1, it returns 1.
}
\examples{
addjust_beta(1.2)  # Will return 1
}
