% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bulk_DNAm}
\alias{bulk_DNAm}
\title{Bulk DNA Methylation Data}
\format{
A matrix with CpGs as rows and samples as columns. Each entry represents
the methylation level of a specific CpG site in a bulk DNA sample.
}
\usage{
bulk_DNAm
}
\description{
This dataset contains bulk DNA methylation (DNAm) data for deconvolution. It represents
the mixed CpG methylation levels from a mixture of different cell types in a biological sample.
}
\examples{
data(bulk_DNAm)
head(bulk_DNAm)
}
\keyword{datasets}
