% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bulk_RNA}
\alias{bulk_RNA}
\title{Bulk RNA-seq Data}
\format{
A matrix with genes as rows and samples as columns. Each entry represents
the expression level of a specific gene in a bulk RNA sample.
}
\usage{
bulk_RNA
}
\description{
This dataset contains bulk RNA-seq data for deconvolution. It represents the mixed
gene expression levels from a mixture of different cell types in a biological sample.
}
\examples{
data(bulk_RNA)
head(bulk_RNA)
}
\keyword{datasets}
