% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_algo.R
\name{EM_deconv}
\alias{EM_deconv}
\title{EM Deconvolution for DNAm and RNA Data}
\usage{
EM_deconv(
  Y,
  D_ng,
  Y_ref,
  D_ng_ref,
  my_pi,
  bulk_rna,
  ref_rna,
  omega,
  num_iterations = 150,
  true_theta = NULL,
  verbose = TRUE,
  rescale = FALSE,
  gspecific = FALSE,
  theta_g = FALSE,
  stopping = 1e-04,
  cell_size
)
}
\arguments{
\item{Y}{Bulk DNAm methylated count data.}

\item{D_ng}{Bulk coverage data.}

\item{Y_ref}{Reference DNAm methylated count data.}

\item{D_ng_ref}{Reference coverage data.}

\item{my_pi}{Initial Pi matrix.}

\item{bulk_rna}{Bulk RNA data.}

\item{ref_rna}{Reference RNA data.}

\item{omega}{Weighting factor between observed and reference data.}

\item{num_iterations}{Maximum number of iterations for the EM algorithm.}

\item{true_theta}{True cell-type proportions (if available, for validation).}

\item{verbose}{Whether to print progress updates.}

\item{rescale}{Whether to rescale data.}

\item{gspecific}{Whether to rescale D_ng and Y gene-specifically.}

\item{theta_g}{Whether to use the same D_ng for all genes during theta updates.}

\item{stopping}{Stopping criterion for theta updates.}

\item{cell_size}{A normalization factor for cell sizes.}
}
\value{
A list containing the updated theta and performance metrics.
}
\description{
This function performs deconvolution using the Expectation-Maximization (EM) algorithm. 
It updates the parameters based on input DNAm and RNA data.
}
\examples{
result <- EM_deconv(Y, D_ng, Y_ref, D_ng_ref, my_pi, omega=0.5, num_iterations=150, true_theta=true_theta_matrix)
}
