% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMixed.R
\name{EMixed}
\alias{EMixed}
\title{EMixed Deconvolution Wrapper for RNA and DNAm Data}
\usage{
EMixed(bulk_RNA, ref_RNA, bulk_DNAm, ref_DNAm, parallel, ncore = 2)
}
\arguments{
\item{bulk_RNA}{Bulk RNA-seq data.}

\item{ref_RNA}{Reference RNA-seq data.}

\item{bulk_DNAm}{Bulk DNA methylation data.}

\item{ref_DNAm}{Reference DNA methylation data.}

\item{parallel}{Logical flag indicating whether to enable parallel computation. If TRUE, parallel computation will be used.}

\item{ncore}{Number of cores to use for parallel computation (default is 2).}
}
\value{
A list containing the averaged deconvolution results for RNA and DNAm.
}
\description{
This function wraps around the EM_wrap function to perform deconvolution using both RNA and DNAm data. It
processes the bulk RNA and DNAm data, ensures there are no rows with all zero values, and calls the EM_wrap
function to perform the deconvolution for each data type (RNA and DNAm). The final estimates are averaged.
}
