% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_algo.R, R/EM_wrap.R
\name{impute01}
\alias{impute01}
\title{Impute 0 and 1 Values with Second Smallest/Largest Unique Values}
\usage{
impute01(data)

impute01(data)
}
\arguments{
\item{data}{A vector or matrix containing the data with 0s and 1s to be replaced.}
}
\value{
A vector or matrix with imputed values.

Returns the modified dataset where 0s and 1s are imputed.
}
\description{
This function replaces 0s in the data with the second smallest unique value, 
and replaces 1s with the second largest unique value.

This function replaces occurrences of 1 with the second-largest unique value and
occurrences of 0 with the second-smallest unique value in the dataset.
}
\examples{
impute01(c(0, 1, 2, 3, 4, 5))  # Example usage
impute01(c(0, 1, 2, 3, 4, 5))  # Example usage
}
