% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_DNAm}
\alias{ref_DNAm}
\title{Reference DNA Methylation Data}
\format{
A matrix with CpGs as rows and cell types as columns. Each entry represents
the beta value of a specific CpG site in a particular cell type.
}
\usage{
ref_DNAm
}
\description{
This dataset contains reference DNA methylation (DNAm) data for deconvolution.
It includes beta value across different cell types, which are used
as a reference for the EMixed deconvolution process.
}
\examples{
data(ref_DNAm)
head(ref_DNAm)
}
\keyword{datasets}
