% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_RNA}
\alias{ref_RNA}
\title{Reference RNA-seq Data}
\format{
A matrix with genes as rows and cell types as columns. Each entry represents
the expression level of a specific gene in a particular cell type.
}
\usage{
ref_RNA
}
\description{
This dataset contains reference RNA-seq data for deconvolution.
It includes gene expression levels across different cell types, which are used
as a reference for the EMixed deconvolution process.
}
\examples{
data(ref_RNA)
head(ref_RNA)
}
\keyword{datasets}
