% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_algo.R
\name{sum_to_one}
\alias{sum_to_one}
\title{Normalize Matrix Rows to Sum to One}
\usage{
sum_to_one(matrix)
}
\arguments{
\item{matrix}{A numeric matrix to be normalized.}
}
\value{
The matrix with rows normalized to sum to 1.
}
\description{
This function ensures each row of the input matrix sums to 1. It handles missing and invalid values appropriately.
}
